Check directory exists before deleting

+++ extensions/ql-vscode/src/vscode-tests/cli-integration/remote-queries/variant-analysis-results-manager.test.ts
@@ -63,7 +63,9 @@ describe(VariantAnalysisResultsManager.name, function() {
     });
 
     afterEach(async () => {
-      fs.rmSync(variantAnalysisStoragePath, { recursive: true });
+      if (fs.existsSync(variantAnalysisStoragePath)) {
+        fs.rmSync(variantAnalysisStoragePath, { recursive: true });
+      }
     });
 
     describe('isVariantAnalysisRepoDownloaded', () => {
